Extract separate class to save page HTML to disk
