+++ src/main/java/nl/hsac/fitnesse/fixture/Environment.java
@@ -557,6 +557,41 @@ public class Environment {
     }
 
     /**
+     * Converts a file path into a relative wiki path, if the path is insides the wiki's 'files' section.
+     * @param filePath path to file.
+     * @return relative URL pointing to the file (so a hyperlink to it can be created).
+     */
+    public String getWikiUrl(String filePath) {
+        String wikiUrl = null;
+        String filesDir = getFitNesseFilesSectionDir();
+        if (filePath.startsWith(filesDir)) {
+            String relativeFile = filePath.substring(filesDir.length());
+            relativeFile = relativeFile.replace('\\', '/');
+            wikiUrl = "files" + relativeFile;
+        }
+        return wikiUrl;
+    }
+
+    /**
+     * Gets absolute path from wiki url, if file exists.
+     * @param wikiUrl a relative path that can be used in wiki page, or any file path.
+     * @return absolute path to the target of the url, if such a file exists; null if the target does not exist.
+     */
+    public String getFilePathFromWikiUrl(String wikiUrl) {
+        String url = getHtmlCleaner().getUrl(wikiUrl);
+        File file;
+        if (url.startsWith("files/")) {
+            String relativeFile = url.substring("files".length());
+            relativeFile = relativeFile.replace('/', File.separatorChar);
+            String pathname = getFitNesseFilesSectionDir() + relativeFile;
+            file = new File(pathname);
+        } else {
+            file = new File(url);
+        }
+        return file.exists() ? file.getAbsolutePath() : url;
+    }
+
+    /**
      * @return default (global) map helper.
      */
     public MapHelper getMapHelper() {
