+++ src/main/java/nl/hsac/fitnesse/fixture/slim/SlimFixture.java
@@ -116,13 +116,7 @@ public class SlimFixture  implements InteractionAwareFixture {
      * @return relative URL pointing to the file (so a hyperlink to it can be created).
      */
     protected String getWikiUrl(String filePath) {
-        String wikiUrl = null;
-        if (filePath.startsWith(filesDir)) {
-            String relativeFile = filePath.substring(filesDir.length());
-            relativeFile = relativeFile.replace('\\', '/');
-            wikiUrl = "files" + relativeFile;
-        }
-        return wikiUrl;
+        return getEnvironment().getWikiUrl(filePath);
     }
 
     /**
@@ -131,17 +125,7 @@ public class SlimFixture  implements InteractionAwareFixture {
      * @return absolute path to the target of the url, if such a file exists; null if the target does not exist.
      */
     protected String getFilePathFromWikiUrl(String wikiUrl) {
-        String url = getUrl(wikiUrl);
-        File file;
-        if (url.startsWith("files/")) {
-            String relativeFile = url.substring("files".length());
-            relativeFile = relativeFile.replace('/', File.separatorChar);
-            String pathname = filesDir + relativeFile;
-            file = new File(pathname);
-        } else {
-            file = new File(url);
-        }
-        return file.exists() ? file.getAbsolutePath() : url;
+        return getEnvironment().getFilePathFromWikiUrl(wikiUrl);
     }
 
 }
